Alterations to pretty log alias
No point typing in the 2, so renamed to lg; also changed the default colours to RGB colours so that different colour profiles on the terminal don't mess with the colouring of the output.  This involves some highly suspect messing about with quotes that I couldn't really find any documentation for.